text_tensor_dedicom/requirements.txt
beautifulsoup4==4.9.1
flair
matplotlib==3.2.2
nltk==3.5
numpy==1.18.4
pandas==1.0.3
Pillow==7.2.0
PyYAML==5.3.1
requests==2.22.0
scipy==1.5.0
scikit-learn==0.23.1
seaborn==0.10.1
selenium==3.141.0
somajo
tensorboard==2.2.2
torch==1.5.1
torchvision==0.6.1
tqdm==4.42.1
transformers
umap-learn
wikipedia-api
